add user_test program

+++ os/lib/system_call.c
@@ -3,6 +3,6 @@
 extern u16 system_call[MAX_SYSTEM_CALL];
 
 void install_system_call(int system_call_number, void (*function)() ){
-    system_call[system_call_number] = function;
+    system_call[system_call_number] = (u16) function;
     return ;
 }
